// src/app/models/admin-models/model.technicalCharacteristics.ts
export class TechnicalCharacteristics
{
    is_publish?: number;
    id?: number;

    ua_type_desc?: string;
    eng_type_desc?: string;
    rus_type_desc?: string;

    name?: string;

    ua_description?: string;
    eng_description?: string;
    rus_description?: string;


    photo?: string;
    tech_photo?:string;

    ua_type_machine?: string;
    eng_type_machine?: string;
    rus_type_machine?: string;


    ua_productivity?: string;
    eng_productivity?: string;
    rus_productivity?: string;

    motor?: string;
    power?: number;
    idle_speed?: number;
    working_speed?: number;


    ua_type_node?: string;
    eng_type_node?: string;
    rus_type_node?:string;


    automatic_no_material?: number;
    shaft_diameter?: number;
    lenght_work?: number;
    height_work?: number;
    width_work?: number;
    lenght_transport?: number;
    height_transport?: number;
    width_transport?: number;
    volume_hydraulic?: number;
    max_pressure_hydraulic?: number;
    size_sieve_holes?: number;
    size_org_fraction?: number;
    interest_fraction_out?:number;
    product_mass?: number;


    ua_desc_title1?: string;
    eng_desc_title1?: string;
    rus_desc_title1?: string;


    ua_desc_title2?: string;
    eng_desc_title2?: string;
    rus_desc_title2?: string;



    ua_desc_title3?: string;
    eng_desc_title3?: string;
    rus_desc_title3?:string;

    
    ua_desc_title4?: string;
    eng_desc_title4?: string;
    rus_desc_title4?: string;



    ua_desc_title5?: string;
    eng_desc_title5?: string;
    rus_desc_title5?: string;



    ua_desc_under_title1?: string;
    eng_desc_under_title1?: string;
    rus_desc_under_title1?: string;


    ua_desc_under_title2?: string;
    eng_desc_under_title2?:string;
    rus_desc_under_title2?: string;


    ua_desc_under_title3?: string;
    eng_desc_under_title3?:string;
    rus_desc_under_title3?: string;

    ua_desc_under_title4?: string;
    eng_desc_under_title4?:string;
    rus_desc_under_title4?: string;

    ua_desc_under_title5?: string;
    eng_desc_under_title5?:string;
    rus_desc_under_title5?: string;

    ua_desc_description1?: string;
    eng_desc_description1?:string;
    rus_desc_description1?: string;

    ua_desc_description2?: string;
    eng_desc_description2?:string;
    rus_desc_description2?: string;

    ua_desc_description3?: string;
    eng_desc_description3?:string;
    rus_desc_description3?: string;

    ua_desc_description4?: string;
    eng_desc_description4?:string;
    rus_desc_description4?: string;

    ua_desc_description5?: string;
    eng_desc_description5?:string;
    rus_desc_description5?: string;


    raw_id1?: number;
    raw_id2?:number;
    raw_id3?: number;

    photo_gallery1?: string;
    photo_gallery2?:string;
    photo_gallery3?: string;
    photo_gallery4?: string;
    photo_gallery5?:string;
    photo_gallery6?: string;
    photo_gallery7?: string;
    photo_gallery8?:string;

    video_url:string;
    
    rating?: number;

    pdf?:string;

    createdAt: string;
    updatedAt: string;
}